refactor(package-name): change package name from data_lunch_app to dlunch
select a shorter and simpler name for PyPI
data_lunch_app is converted to dlunch
project name is set to dlunch
PANEL_APP env variable and docker image name are left as data_lunch_app
