+++ MANIFEST.in
@@ -1,4 +1,4 @@
-recursive-include data_lunch_app/conf *
-recursive-include data_lunch_app/static *
-recursive-include data_lunch_app/templates *
-include data_lunch_app/quotes.xlsx
+recursive-include dlunch/conf *
+recursive-include dlunch/static *
+recursive-include dlunch/templates *
+include dlunch/quotes.xlsx
