+++ dlunch/conf/basic_auth/basic_auth.yaml
@@ -3,7 +3,7 @@
 
 # AUTH PROVIDER OBJECT
 auth_provider:
-  _target_: data_lunch_app.auth.DataLunchProvider
+  _target_: dlunch.auth.DataLunchProvider
   login_template: ${package_path}/templates/login_basic.html
   logout_template: ${package_path}/templates/logout.html
 # Those values are  used by basic authentication specific features
