+++ dlunch/conf/panel/default.yaml
@@ -82,7 +82,7 @@ scheduled_tasks: # Set to [] empty list to turn it off
       minute: null
       period: 1d
     callable:
-      _target_: data_lunch_app.scheduled_tasks.reset_guest_user_password
+      _target_: dlunch.scheduled_tasks.reset_guest_user_password
   - kwargs:
       name: scheduled cleaning
       enabled: true
@@ -90,7 +90,7 @@ scheduled_tasks: # Set to [] empty list to turn it off
       minute: 00
       period: 1d
     callable:
-      _target_: data_lunch_app.scheduled_tasks.clean_files_db
+      _target_: dlunch.scheduled_tasks.clean_files_db
   - kwargs:
       name: database upload
       enabled: ${db.ext_storage_upload.enabled}
@@ -98,7 +98,7 @@ scheduled_tasks: # Set to [] empty list to turn it off
       minute: 25
       period: 60min
     callable:
-      _target_: data_lunch_app.scheduled_tasks.upload_db_to_gcp_storage
+      _target_: dlunch.scheduled_tasks.upload_db_to_gcp_storage
       source_file_name: ${db.ext_storage_upload.source_file_name}
       destination_blob_name: ${db.ext_storage_upload.destination_blob_name}
       bucket_name: ${db.ext_storage_upload.bucket_name}
