+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "data-lunch"
+name = "dlunch"
 version = "2.7.0"
 authors = [
   { name="Michele Alberti", email="michele.alberti90@gmail.com" },
@@ -25,10 +25,10 @@ Homepage = "https://github.com/Michele-Alberti/data-lunch"
 Issues = "https://github.com/Michele-Alberti/data-lunch/issues"
 
 [project.scripts]
-data-lunch = "data_lunch_app.cli:main"
+data-lunch = "dlunch.cli:main"
 
 [tool.setuptools]
-packages = ["data_lunch_app"]
+packages = ["dlunch"]
 
 [tool.setuptools.dynamic]
 dependencies = {file = ["requirements/requirements.txt"]}
@@ -54,7 +54,7 @@ exclude = '''
 name = "cz_conventional_commits"
 version = "2.7.0"
 version_files = [
-    "data_lunch_app/__init__.py",
+    "dlunch/__init__.py",
     "pyproject.toml:version",
 ]
 tag_format = "v$version"
